fix some bug

+++ include/collection.class.php
@@ -37,11 +37,12 @@ class collection{
         }
         return $del;
     }
-    public static function order($id,$name,$phone,$address,$delivery,$product_id,$product_name,$img,$color,$size,$quantity,$price){
+    public static function order($id,$name,$phone,$address,$delivery,$product_id,$size){
         $conn = DB::db();
         $status1="confirm";
         $status2="pending";
-        $query="SELECT * FROM checkout WHERE user_id='$id' AND product_id='$product_id' AND size='$size' AND status='$status2' OR status='$status1'";
+        $order_id="ORDER_".rand(1111,9999).substr($phone,8);
+        $query="SELECT * FROM checkout WHERE user_id='$id' AND product_id IN('$product_id') AND size IN('$size') AND status='$status2' OR status='$status1'";
         $result=$conn->query($query);
         date_default_timezone_set('Asia/Kolkata');
         $current_date=date("Y/m/d H:i:s");
@@ -50,47 +51,61 @@ class collection{
             // header("refresh:1; url=order");
             $order=false;
         }else{
-            $order_id="ORDER_".substr($phone,5);
-            $insert="INSERT INTO checkout (user_id,order_id,name,phone,address,delivery,product_id,product_name,product_img,size,quantity,color,price,time) VALUES('$id','$order_id','$name','$phone','$address','$delivery','$product_id','$product_name','$img','$size','$quantity','$color','$price','$current_date')";
-            if($conn->query($insert)===TRUE){
-                $update="DELETE FROM cart WHERE user_id='$id' AND product_id='$product_id' AND size='$size'";
-                if($conn->query($update)===TRUE){
-                    $sql=mysqli_query($conn,"SELECT quantity FROM mens_size WHERE product_id='$product_id' AND size='$size'");
-                    $get=mysqli_fetch_array($sql);
-                    $qty=$get["quantity"]-$quantity;
-                    $change=mysqli_query($conn,"UPDATE mens_size SET quantity='$qty' WHERE product_id='$product_id' AND size='$size'");
-                    if($change===true){
-                        $order=true;
-                    }else{
-                        $order=false;
-                        $order=$conn->error;
+            $place="INSERT INTO orders (user_id,order_id,name,phone,address,delivery,time) VALUES('$id','$order_id','$name','$phone','$address','$delivery','$current_date')";
+            if($conn->query($place)===TRUE){
+                $sql=mysqli_query($conn,"SELECT * FROM cart WHERE user_id='$id'");
+                while($res=mysqli_fetch_array($sql)){
+                    $product_id1=$res["product_id"];
+                    $product_name=$res["product_name"];
+                    $img=$res["product_img"];
+                    $size1=$res["size"];
+                    $quantity=$res["quantity"];
+                    $color=$res["color"];
+                    $total=$res["total"];
+                    $insert="INSERT INTO checkout (order_id,user_id,product_id,product_name,product_img,size,quantity,color,price) VALUES ('$order_id','$id','$product_id1','$product_name','$img','$size1','$quantity','$color','$total')";
+                    if($conn->query($insert)===true){
+                        $update="DELETE FROM cart WHERE user_id='$id' AND product_id='$product_id1' AND size='$size1'";
+                        if($conn->query($update)===TRUE){
+                            $sql1=mysqli_query($conn,"SELECT quantity FROM mens_size WHERE product_id='$product_id1' AND size='$size1'");
+                            $get=mysqli_fetch_array($sql1);
+                            $qty=$get["quantity"]-$quantity;
+                            $change=mysqli_query($conn,"UPDATE mens_size SET quantity='$qty' WHERE product_id='$product_id1' AND size='$size1'");
+                            if($change===true){
+                                $order=true;
+                            }else{
+                                $order=false;
+                                $order=$conn->error;
+                            }
+                        }else{
+                            $order=false;
+                            $order=$conn->error;
+                        }
                     }
-                }else{
-                    $order=false;
-                    $order=$conn->error;
                 }
+
             }else{
                 $order=false;
                 $order=$conn->error;
             }
         }
+                   
         return $order;
     }
     public static function remove_order($id){
         $conn = DB::db();
-        $remove="UPDATE checkout SET status='cancel' WHERE id='$id'";
-        if($conn->query($remove)===TRUE){
-            $get=mysqli_query($conn,"SELECT * FROM checkout WHERE id='$id'");
-            $fetch=mysqli_fetch_array($get);
-            $product_id=$fetch["product_id"];
-            $size=$fetch["size"];
-            $quantity=$fetch["quantity"];
-            $sql=mysqli_query($conn,"SELECT quantity FROM mens_size WHERE product_id='$product_id' AND size='$size'");
-            $result=mysqli_fetch_array($sql);
-            $quantity1=$result["quantity"];
-            $qty=$quantity+$quantity1;
-            $update=mysqli_query($conn,"UPDATE mens_size SET quantity='$qty' WHERE product_id='$product_id' AND size='$size'");
-            if($update===true){
+        $get=mysqli_query($conn,"SELECT * FROM checkout WHERE id='$id'");
+        $fetch=mysqli_fetch_array($get);
+        $product_id=$fetch["product_id"];
+        $size=$fetch["size"];
+        $quantity=$fetch["quantity"];
+        $sql=mysqli_query($conn,"SELECT quantity FROM mens_size WHERE product_id='$product_id' AND size='$size'");
+        $result=mysqli_fetch_array($sql);
+        $quantity1=$result["quantity"];
+        $qty=$quantity+$quantity1;
+        $update=mysqli_query($conn,"UPDATE mens_size SET quantity='$qty' WHERE product_id='$product_id' AND size='$size'");
+        if($update===true){
+            $remove="UPDATE checkout SET status='remove' WHERE id='$id'";
+            if($conn->query($remove)===TRUE){
                 $del=true;
             }else{
                 $del=false;
@@ -106,26 +121,32 @@ class collection{
         $conn = DB::db();
         $status1="confirm";
         $status2="pending";
+        $order_id="ORDER_".rand(1111,9999).substr($phone,8);
         $query="SELECT * FROM checkout WHERE user_id='$user_id' AND product_id='$product_id' AND size='$size' AND status='$status2' OR status='$status1'";
         $result=$conn->query($query);
         date_default_timezone_set('Asia/Kolkata');
-        $current_date=date("Y/m/d");
+        $current_date=date("Y/m/d H:i:s");
         if($result->num_rows>0){
             echo"<script>alert('This product already placed in order please check your order list!')</script>";
             header("refresh:1; url=order");
             $check=false;
         }else{
-            $order_id="ORDER_".substr($phone,5);
-            $insert="INSERT INTO checkout (user_id,order_id,name,phone,address,delivery,product_id,product_name,product_img,size,quantity,color,price,time) VALUES('$user_id','$order_id','$name','$phone','$address','$delivery','$product_id','$product_name','$img','$size','$quantity','$color','$total','$current_date')";
-            if($conn->query($insert)===TRUE){
-                $update="DELETE FROM cart WHERE user_id='$user_id' AND product_id='$product_id' AND size='$size'";
-                if($conn->query($update)===TRUE){
-                    $sql=mysqli_query($conn,"SELECT quantity FROM mens_size WHERE product_id='$product_id' AND size='$size'");
-                    $get=mysqli_fetch_array($sql);
-                    $qty=$get["quantity"]-$quantity;
-                    $change=mysqli_query($conn,"UPDATE mens_size SET quantity='$qty' WHERE product_id='$product_id' AND size='$size'");
-                    if($change===true){
-                        $check=true;
+            $place="INSERT INTO orders (user_id,order_id,name,phone,address,delivery,time) VALUES('$user_id','$order_id','$name','$phone','$address','$delivery','$current_date')";
+            if($conn->query($place)===TRUE){
+                $insert="INSERT INTO checkout (order_id,user_id,product_id,product_name,product_img,size,quantity,color,price) VALUES('$order_id','$user_id','$product_id','$product_name','$img','$size','$quantity','$color','$total')";
+                if($conn->query($insert)===TRUE){     
+                    $update="DELETE FROM cart WHERE user_id='$user_id' AND product_id='$product_id' AND size='$size'";
+                    if($conn->query($update)===TRUE){
+                        $sql=mysqli_query($conn,"SELECT quantity FROM mens_size WHERE product_id='$product_id' AND size='$size'");
+                        $get=mysqli_fetch_array($sql);
+                        $qty=$get["quantity"]-$quantity;
+                        $change=mysqli_query($conn,"UPDATE mens_size SET quantity='$qty' WHERE product_id='$product_id' AND size='$size'");
+                        if($change===true){
+                            $check=true;
+                        }else{
+                            $check=false;
+                            $check=$conn->error;
+                        }
                     }else{
                         $check=false;
                         $check=$conn->error;
@@ -134,12 +155,11 @@ class collection{
                     $check=false;
                     $check=$conn->error;
                 }
+                    
             }else{
                 $check=false;
                 $check=$conn->error;
-            }
-                    
-                
+            }    
         }
         return $check;
     }
